Edit submit solution endpoint and gat all problems endpoint validation

+++ CodeSphere.Application/Features/Problem/Queries/GetAll/GetAllQueryHandler.cs
@@ -1,17 +1,9 @@
 ﻿using AutoMapper;
 using CodeSphere.Domain.Abstractions;
-using CodeSphere.Domain.Abstractions.Repositories;
-using CodeSphere.Domain.Models.Entities;
 using CodeSphere.Domain.Premitives;
 using MediatR;
 using Microsoft.IdentityModel.Tokens;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
-using System.Reflection.Metadata.Ecma335;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace CodeSphere.Application.Features.Problem.Queries.GetAll
 {
@@ -29,9 +21,9 @@ namespace CodeSphere.Application.Features.Problem.Queries.GetAll
 
         public async Task<Response> Handle(GetAllProblemsQuery request, CancellationToken cancellationToken)
         {
-            var problems = await _unitOfWork.ElasticSearchRepository.SearchProblemsAsync(request.ProblemName, request.TopicsIds,(int) request.Difficulty);
-            
-            if(problems.IsNullOrEmpty())
+            var problems = await _unitOfWork.ElasticSearchRepository.SearchProblemsAsync(request.ProblemName, request.TopicsIds, request.Difficulty);
+
+            if (problems.IsNullOrEmpty())
             {
                 return await Response.FailureAsync("No Problems Found", HttpStatusCode.NotFound);
             }
@@ -40,7 +32,7 @@ namespace CodeSphere.Application.Features.Problem.Queries.GetAll
 
             foreach (var problem in problems)
             {
-                var submission =  _unitOfWork.SubmissionRepository.GetSolvedSubmissions(problem.Id, request.UserId);
+                var submission = _unitOfWork.SubmissionRepository.GetSolvedSubmissions(problem.Id, request.UserId);
                 var x = _mapper.Map<GetAllQueryResponse>(problem);
                 x.IsSolved = !submission.IsNullOrEmpty();
                 responses.Add(x);
